# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Memory API",
    "description": "Private memory and project API. Prefer runAssistantCommand for GPT use; it handles commands, active project context, tasks, notes, meetings, memories, search, summaries, and delete.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://memory-green-kappa.vercel.app"
    }
  ],
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
    "/api/assistant/command": {
      "post": {
        "operationId": "runAssistantCommand",
        "summary": "Run assistant command",
        "description": "Run one command or natural language request. Handles active project, tasks, notes, meetings, memories, search, and delete.",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": false,
            "description": "Optional session key for active project context.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assistant command result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantCommandResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "AssistantCommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "type": "string",
            "description": "User command or natural language request.",
            "examples": [
              "@project HRMS",
              "@task Implement semantic search",
              "@memory My passport expires in 2032",
              "@find jwt",
              "work on hrms"
            ],
            "minLength": 1
          },
          "sessionId": {
            "type": "string",
            "description": "Optional session key for active project context."
          }
        }
      },
      "AssistantCommandResponse": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "activeProject": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "needsProject": {
            "type": "boolean"
          },
          "needsSelection": {
            "type": "boolean"
          },
          "data": {
            "description": "Command-specific result data."
          }
        }
      },
...
      "ProjectSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "_id",
          "name"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}